rename maru to example
