+++ android/src/main/AndroidManifest.xml
@@ -1,10 +1,10 @@
 <?xml version="1.0" encoding="utf-8"?>
 <manifest xmlns:android="http://schemas.android.com/apk/res/android"
-    package="com.maru.twitter_login">
+    package="com.example.twitter_login">
     <application>
 
         <activity
-            android:name="com.maru.twitter_login.chrome_custom_tabs.ChromeCustomTabsActivity"/>
+            android:name="com.example.twitter_login.chrome_custom_tabs.ChromeCustomTabsActivity"/>
     </application>
 
     <queries>
