+++ android/src/main/kotlin/com/example/twitter_login/chrome_custom_tabs/ChromeCustomTabsActivity.java
@@ -1,4 +1,4 @@
-package com.maru.twitter_login.chrome_custom_tabs;
+package com.example.twitter_login.chrome_custom_tabs;
 
 import android.app.Activity;
 import android.content.Intent;
@@ -8,8 +8,8 @@ import android.os.Bundle;
 import androidx.browser.customtabs.CustomTabsIntent;
 import androidx.browser.customtabs.CustomTabsSession;
 
-import com.maru.twitter_login.R;
-import com.maru.twitter_login.customtabsclient.CustomTabsHelper;
+import com.example.twitter_login.R;
+import com.example.twitter_login.customtabsclient.CustomTabsHelper;
 
 import org.jetbrains.annotations.NotNull;
 
@@ -64,8 +64,7 @@ public class ChromeCustomTabsActivity extends Activity implements MethodChannel.
                         chromeCustomTabsActivity,
                         customTabsIntent,
                         uri,
-                        CHROME_CUSTOM_TAB_REQUEST_CODE
-                );
+                        CHROME_CUSTOM_TAB_REQUEST_CODE);
             }
 
             @Override
@@ -80,7 +79,6 @@ public class ChromeCustomTabsActivity extends Activity implements MethodChannel.
     public void onMethodCall(final MethodCall call, @NotNull final MethodChannel.Result result) {
         // invokeMethod is defined to pass the cancelled operation to the Dart side.
     }
-
 
     private void prepareCustomTabsIntent(CustomTabsIntent customTabsIntent) {
         customTabsIntent.intent.setPackage(CustomTabsHelper.getPackageNameToUse(this));
