+++ android/src/main/kotlin/com/example/twitter_login/customtabsclient/ServiceConnectionCallback.java
@@ -1,11 +1,6 @@
-package com.maru.twitter_login.customtabsclient;
+package com.example.twitter_login.customtabsclient;
 
-import android.app.Service;
-import android.content.Intent;
-import android.os.Binder;
-import android.os.IBinder;
-
-import androidx.annotation.Nullable;
+import androidx.browser.customtabs.CustomTabsClient;
 
 // Copyright 2015 Google Inc. All Rights Reserved.
 //
@@ -20,12 +15,21 @@ import androidx.annotation.Nullable;
 // WITHOUT WARRANTIES OR CONDITIONS OF ANY KIND, either express or implied.
 // See the License for the specific language governing permissions and
 // limitations under the License.
-public class KeepAliveService extends Service {
-    private static final Binder sBinder = new Binder();
 
-    @Nullable
-    @Override
-    public IBinder onBind(Intent intent) {
-        return sBinder;
-    }
+/**
+ * Callback for events when connecting and disconnecting from Custom Tabs
+ * Service.
+ */
+public interface ServiceConnectionCallback {
+    /**
+     * Called when the service is connected.
+     * 
+     * @param client a CustomTabsClient
+     */
+    void onServiceConnected(CustomTabsClient client);
+
+    /**
+     * Called when the service is disconnected.
+     */
+    void onServiceDisconnected();
 }
